feat: implement get user top endpoint
