+++ api/gateway/native/internal/entity/book.go
@@ -45,85 +45,3 @@ func (bs Books) Map() map[int64]*Book {
 	}
 	return res
 }
-
-type Bookshelf struct {
-	*pb.Bookshelf
-}
-
-type Bookshelves []*Bookshelf
-
-func NewBookshelf(b *pb.Bookshelf) *Bookshelf {
-	return &Bookshelf{b}
-}
-
-func (b *Bookshelf) Status() BookshelfStatus {
-	return NewBookshelfStatus(b.GetStatus())
-}
-
-func NewBookshelves(bs []*pb.Bookshelf) Bookshelves {
-	res := make(Bookshelves, len(bs))
-	for i := range bs {
-		res[i] = NewBookshelf(bs[i])
-	}
-	return res
-}
-
-func (bs Bookshelves) BookIDs() []int64 {
-	bookIDs := []int64{}
-	books := map[int64]bool{}
-	for _, b := range bs {
-		if _, ok := books[b.BookId]; ok {
-			continue
-		}
-
-		books[b.BookId] = true
-		bookIDs = append(bookIDs, b.BookId)
-	}
-	return bookIDs
-}
-
-type Review struct {
-	*pb.Review
-}
-
-type Reviews []*Review
-
-func NewReview(r *pb.Review) *Review {
-	return &Review{r}
-}
-
-func NewReviews(rs []*pb.Review) Reviews {
-	res := make(Reviews, len(rs))
-	for i := range rs {
-		res[i] = NewReview(rs[i])
-	}
-	return res
-}
-
-func (rs Reviews) UserIDs() []string {
-	userIDs := []string{}
-	reviews := map[string]bool{}
-	for _, r := range rs {
-		if _, ok := reviews[r.UserId]; ok {
-			continue
-		}
-
-		reviews[r.UserId] = true
-		userIDs = append(userIDs, r.UserId)
-	}
-	return userIDs
-}
-
-func (rs Reviews) BookIDs() []int64 {
-	bookIDs := []int64{}
-	reviews := map[int64]bool{}
-	for _, r := range rs {
-		if _, ok := reviews[r.BookId]; ok {
-			continue
-		}
-
-		reviews[r.BookId] = true
-		bookIDs = append(bookIDs, r.BookId)
-	}
-	return bookIDs
-}
