+++ api/gateway/native/internal/entity/top_test.go
@@ -1,0 +1,1 @@
+package entity
